Refactor table styles and improve pagination layout; update script paths for user management
