+++ src/assets/js/pages/roles-permissions/groups/table.js
@@ -1,4 +1,4 @@
-import { getPathPrefix } from "../../../../js/path-utils.js";
+import { getPathPrefix } from "../../../path-utils.js";
 ("use strict")
 
 // Class definition
@@ -72,7 +72,7 @@ var GroupsDatatable = (function () {
                       </button>
                       <ul class="dropdown-menu">
                         <li><a class="dropdown-item d-flex align-items-center gap-2" href="` +
-              getPathPrefix("/roles_permissions/groups/edit") +
+              getPathPrefix("/roles-permissions/groups/edit") +
               `"><i class="ri-pencil-line"></i> Edit</a></li>
                         <li><hr class="dropdown-divider"></li>
                         <li><a class="dropdown-item d-flex align-items-center gap-2 text-danger delete-button" href="#"><i class="ri-delete-bin-line"></i> Delete</a></li>
@@ -119,6 +119,7 @@ var GroupsDatatable = (function () {
 
     // Filter datatable on submit
     filterButton.addEventListener("click", function () {
+      $(".dataTables_wrapper").addClass('processing')
       $(".dataTables_processing").css("display", "block")
       $.fn.dataTable.ext.search = []
       dt.draw()
@@ -161,6 +162,7 @@ var GroupsDatatable = (function () {
     // Simulate loading delay
     setTimeout(function () {
       dt.draw()
+      $(".dataTables_wrapper").removeClass('processing')
       $(".dataTables_processing").css("display", "none")
     }, 500)
   }
@@ -172,6 +174,7 @@ var GroupsDatatable = (function () {
 
     // Reset datatable
     resetButton.addEventListener("click", function () {
+      $(".dataTables_wrapper").addClass('processing')
       $(".dataTables_processing").css("display", "block")
 
       if ($(".form-check-input:checked").length > 0) {
@@ -186,6 +189,7 @@ var GroupsDatatable = (function () {
       // Simulate loading delay
       setTimeout(function () {
         dt.draw()
+        $(".dataTables_wrapper").removeClass('processing')
         $(".dataTables_processing").css("display", "none")
       }, 500)
 
@@ -267,6 +271,7 @@ var GroupsDatatable = (function () {
           }
         }
 
+        $(".dataTables_wrapper").addClass('processing')
         $(".dataTables_processing").css("display", "block")
         $.fn.dataTable.ext.search = []
         setTimeout(function () {
