+++ src/assets/js/pages/roles-permissions/roles/toggle-permissions.js
