+++ src/assets/js/pages/tables/datatable.js
@@ -170,6 +170,7 @@ var ProductsDatatable = (function () {
     // Filter datatable on submit
     filterButton.addEventListener("click", function () {
       $(".dataTables_processing").css("display", "block")
+      $(".dataTables_wrapper").addClass('processing')
       // Trigger the dismiss click on the close button
       document.querySelector('#filterDrawer [data-bs-dismiss="offcanvas"]').click()
       $.fn.dataTable.ext.search = []
@@ -230,6 +231,7 @@ var ProductsDatatable = (function () {
     // Simulate loading delay
     setTimeout(function () {
       dt.draw()
+      $(".dataTables_wrapper").removeClass('processing')
       $(".dataTables_processing").css("display", "none")
     }, 500)
   }
@@ -242,6 +244,7 @@ var ProductsDatatable = (function () {
     // Reset datatable
     resetButton.addEventListener("click", function () {
       // Reset datatable --- official docs reference: https://datatables.net/reference/api/search()
+      $(".dataTables_wrapper").addClass('processing')
       $(".dataTables_processing").css("display", "block")
       // Trigger the dismiss click on the close button
       document.querySelector('#filterDrawer [data-bs-dismiss="offcanvas"]').click()
@@ -262,6 +265,7 @@ var ProductsDatatable = (function () {
       // Simulate loading delay
       setTimeout(function () {
         dt.draw()
+        $(".dataTables_wrapper").removeClass('processing')
         $(".dataTables_processing").css("display", "none")
       }, 500)
     })
@@ -553,6 +557,7 @@ var ProductsDatatable = (function () {
           }
         }
         // dt.draw();
+        $(".dataTables_wrapper").addClass('processing')
         $(".dataTables_processing").css("display", "block")
         $.fn.dataTable.ext.search = []
         setTimeout(function () {
@@ -568,6 +573,7 @@ var ProductsDatatable = (function () {
     if ($('[data-type="navbar"]').length > 0) {
       $('[data-type="navbar"]').on("click", function (e) {
         e.preventDefault()
+        $(".dataTables_wrapper").addClass('processing')
         $(".dataTables_processing").css("display", "block")
         $('[data-type="navbar"]').removeClass("active border-bottom border-2 border-dark")
         $(this).addClass("active border-bottom border-2 border-dark")
