+++ src/assets/js/pages/users/list.js
@@ -168,6 +168,7 @@ var UsersDatatable = (function () {
 
     // Filter datatable on submit
     filterButton.addEventListener("click", function () {
+      $(".dataTables_wrapper").addClass('processing')
       $(".dataTables_processing").css("display", "block")
       // Trigger the dismiss click on the close button
       document.querySelector('#filterDrawer [data-bs-dismiss="offcanvas"]').click()
@@ -209,6 +210,7 @@ var UsersDatatable = (function () {
     // Simulate loading delay
     setTimeout(function () {
       dt.draw()
+      $(".dataTables_wrapper").removeClass('processing')
       $(".dataTables_processing").css("display", "none")
     }, 500)
   }
@@ -221,6 +223,7 @@ var UsersDatatable = (function () {
     // Reset datatable
     resetButton.addEventListener("click", function () {
       // Reset datatable --- official docs reference: https://datatables.net/reference/api/search()
+      $(".dataTables_wrapper").addClass('processing')
       $(".dataTables_processing").css("display", "block")
       // Trigger the dismiss click on the close button
       document.querySelector('#filterDrawer [data-bs-dismiss="offcanvas"]').click()
@@ -241,6 +244,7 @@ var UsersDatatable = (function () {
       // Simulate loading delay
       setTimeout(function () {
         dt.draw()
+        $(".dataTables_wrapper").removeClass('processing')
         $(".dataTables_processing").css("display", "none")
       }, 500)
     })
@@ -532,6 +536,7 @@ var UsersDatatable = (function () {
           }
         }
         // dt.draw();
+        $(".dataTables_wrapper").addClass('processing')
         $(".dataTables_processing").css("display", "block")
         $.fn.dataTable.ext.search = []
         setTimeout(function () {
@@ -547,6 +552,7 @@ var UsersDatatable = (function () {
     if ($('[data-type="navbar"]').length > 0) {
       $('[data-type="navbar"]').on("click", function (e) {
         e.preventDefault()
+        $(".dataTables_wrapper").addClass('processing')
         $(".dataTables_processing").css("display", "block")
         $('[data-type="navbar"]').removeClass("active border-bottom border-2 border-dark")
         $(this).addClass("active border-bottom border-2 border-dark")
