+++ src/html/pages/roles-permissions/permissions/edit.astro
@@ -4,13 +4,13 @@ import PageHeader from "../../../components/common/page-header.astro";
 import { getAssetPrefix, getPathPrefix } from "../../../../utils/path.js";
 
 const pageHeaderProps = {
-  title: "Add Permission",
-  backLink: getPathPrefix("/roles_permissions/permissions/list"),
+  title: "Edit Permission",
+  backLink: getPathPrefix("/roles-permissions/permissions/list"),
   breadcrumbs: [
     { label: "Dashboard", href: getPathPrefix("/dashboard") },
-    { label: "Roles & Permissions" },
-    { label: "Permissions", href: getPathPrefix("/roles_permissions/permissions/list") },
-    { label: "Add Permission", class: "active" }
+    { label: "Roles & Permissions", href: getPathPrefix("/roles-permissions") },
+    { label: "Permissions", href: getPathPrefix("/roles-permissions/permissions/list") },
+    { label: "Edit Permission", class: "active" }
   ]
 };
 
@@ -96,24 +96,24 @@ const pageScript = `
 
         // Show loading state
         submitBtn.disabled = true;
-        submitBtn.innerHTML = spinner + 'Creating...';
+        submitBtn.innerHTML = spinner + 'Updating...';
 
         try {
           // Simulate API call (replace with your actual API endpoint)
           await new Promise(resolve => setTimeout(resolve, 2000));
 
           // On success
-          window.location.href = '${getPathPrefix("/roles_permissions/permissions/list")}';
+          window.location.href = '${getPathPrefix("/roles-permissions/permissions/list")}';
         } catch (error) {
           // On error
           console.error('Error:', error);
           submitBtn.disabled = false;
-          submitBtn.innerHTML = 'Create Permission';
+          submitBtn.innerHTML = 'Update Permission';
 
           // Show error alert
           const alert = document.createElement('div');
           alert.className = 'alert alert-danger mt-3';
-          alert.textContent = 'An error occurred while creating the permission. Please try again.';
+          alert.textContent = 'An error occurred while updating the permission. Please try again.';
           form.insertBefore(alert, form.firstChild);
 
           // Remove alert after 5 seconds
@@ -133,9 +133,9 @@ const pageScript = `
 ---
 
 <AdminLayout
-  title="Add Permission"
-  description="Create New Permission"
-  currentPath="roles_permissions/permissions/add"
+  title="Edit Permission"
+  description="Edit Permission"
+  currentPath="roles-permissions/permissions/edit"
   pageScript={pageScript}
 >
   <PageHeader {...pageHeaderProps} />
@@ -158,6 +158,7 @@ const pageScript = `
                     id="permissionTitle"
                     name="permissionTitle"
                     placeholder="Enter permission title"
+                    value="manage_users"
                   />
                   <div class="form-text">Example: manage_users, view_dashboard, etc.</div>
                 </div>
@@ -177,6 +178,7 @@ const pageScript = `
                     id="displayName"
                     name="displayName"
                     placeholder="Enter display name"
+                    value="Manage Users"
                   />
                   <div class="form-text">Human-readable name for the permission</div>
                 </div>
@@ -194,7 +196,7 @@ const pageScript = `
                     <option value="">Select Module</option>
                     <option value="CMS">CMS</option>
                     <option value="SEO">SEO</option>
-                    <option value="System">System</option>
+                    <option value="System" selected>System</option>
                   </select>
                 </div>
               </div>
@@ -211,7 +213,7 @@ const pageScript = `
                     <option value="">Select Group</option>
                     <option value="Content Management">Content Management</option>
                     <option value="Media Management">Media Management</option>
-                    <option value="User Management">User Management</option>
+                    <option value="User Management" selected>User Management</option>
                     <option value="System Settings">System Settings</option>
                     <option value="System Maintenance">System Maintenance</option>
                     <option value="System Monitoring">System Monitoring</option>
@@ -228,14 +230,31 @@ const pageScript = `
                 </div>
               </div>
             </div>
+
+            <!-- Status -->
+            <div class="row mb-3">
+              <div class="col-12">
+                <div class="form-group">
+                  <label for="status" class="form-label"
+                    >Status <span class="text-danger">*</span></label
+                  >
+                  <select class="form-select" id="status" name="status" required>
+                    <option value="">Select Status</option>
+                    <option value="Active" selected>Active</option>
+                    <option value="Inactive">Inactive</option>
+                    <option value="Trashed">Trashed</option>
+                  </select>
+                </div>
+              </div>
+            </div>
           </div>
           <div class="card-footer">
             <!-- Form Actions -->
             <div class="d-flex justify-content-end gap-2">
-              <a href={getPathPrefix("/roles_permissions/permissions/list")} class="btn btn-light"
+              <a href={getPathPrefix("/roles-permissions/permissions/list")} class="btn btn-light"
                 >Cancel</a
               >
-              <button type="submit" class="btn btn-primary">Create Permission</button>
+              <button type="submit" class="btn btn-primary">Update Permission</button>
             </div>
           </div>
         </form>
