+++ src/html/pages/tables/advanced.astro
@@ -35,7 +35,7 @@ const advancedTableTemplate = `<div class="astero-table">
         </div>
         <!-- Action Buttons -->
         <div class="d-flex gap-2 flex-wrap">
-            <button class="btn btn-light-primary gap-1" data-bs-toggle="offcanvas" data-bs-target="#filterDrawer" aria-label="Open filters">
+            <button class="btn btn-light-primary gap-1 filter-btn" data-bs-toggle="offcanvas" data-bs-target="#filterDrawer" aria-label="Open filters">
                 <i class="ri-filter-3-line me-1" aria-hidden="true"></i> Filters
             </button>
 
@@ -57,13 +57,13 @@ const advancedTableTemplate = `<div class="astero-table">
                             <input class="form-check-input select-all-checkbox" type="checkbox" id="selectAll" aria-label="Select all orders">
                         </div>
                     </th>
-                    <th scope="col"><i class="ri-file-list-3-line me-1" aria-hidden="true"></i>ORDER ID</th>
-                    <th scope="col"><i class="ri-store-2-line me-1" aria-hidden="true"></i>SHOP</th>
-                    <th scope="col"><i class="ri-user-3-line me-1" aria-hidden="true"></i>CUSTOMER</th>
-                    <th scope="col"><i class="ri-money-dollar-circle-line me-1" aria-hidden="true"></i>PRICE</th>
-                    <th scope="col"><i class="ri-calendar-2-line me-1" aria-hidden="true"></i>DATE</th>
-                    <th scope="col"><i class="ri-flag-2-line me-1" aria-hidden="true"></i>STATUS</th>
-                    <th class="text-end" scope="col"><i class="ri-settings-3-line me-1" aria-hidden="true"></i>ACTIONS</th>
+                    <th scope="col">ORDER ID</th>
+                    <th scope="col">SHOP</th>
+                    <th scope="col">CUSTOMER</th>
+                    <th scope="col">PRICE</th>
+                    <th scope="col">DATE</th>
+                    <th scope="col">STATUS</th>
+                    <th class="pe-3 text-end" scope="col">ACTIONS</th>
                 </tr>
             </thead>
             <tbody>
