+++ src/html/pages/tables/datatable.astro
@@ -183,7 +183,7 @@ const productsTableTemplate = `<div class="astero-table">
             <table class="table table-hover table-row-dashed mb-0" id="table_products" aria-label="Products management table">
                 <thead>
                     <tr>
-                        <th class="checkbox-column" width="20" scope="col">
+                        <th class="checkbox-column fs-6" width="20" scope="col">
                             <div class="form-check">
                                 <input class="form-check-input select-all-checkbox" type="checkbox" data-table-select="select_all" aria-label="Select all products">
                             </div>
@@ -193,7 +193,7 @@ const productsTableTemplate = `<div class="astero-table">
                         <th scope="col"><i class="ri-stack-line me-1" aria-hidden="true"></i>STOCK</th>
                         <th scope="col"><i class="ri-money-dollar-circle-line me-1" aria-hidden="true"></i>PRICE</th>
                         <th scope="col"><i class="ri-flag-2-line me-1" aria-hidden="true"></i>STATUS</th>
-                        <th class="text-end" scope="col"><i class="ri-settings-3-line me-1" aria-hidden="true"></i>ACTIONS</th>
+                        <th class="text-end pe-3" scope="col"><i class="ri-settings-3-line me-1" aria-hidden="true"></i>ACTIONS</th>
                     </tr>
                 </thead>
                 <tbody>
