+++ src/html/pages/users/edit.astro
@@ -118,7 +118,7 @@ const permissionModules = [
 
 // External Dependencies: Required JavaScript
 const userFormScripts = `
-  <script src="${getAssetPrefix("/assets/js/pages/roles_and_permissions/roles/toggle-permissions.js")}"></script>
+  <script src="${getAssetPrefix("/assets/js/pages/roles-permissions/roles/toggle-permissions.js")}"></script>
   <script type="module" src="${getAssetPrefix("/assets/js/pages/users/form.js")}"></script>
 `;
