+++ src/html/pages/users/list.astro
@@ -277,7 +277,7 @@ const userListHeaderConfig = {
                 <table class="table table-row-dashed mb-0" id="table_users">
                   <thead>
                     <tr>
-                      <th class="checkbox-column" style="width: 20px">
+                      <th class="checkbox-column fs-6" style="width: 20px">
                         <div class="form-check">
                           <input
                             class="form-check-input select-all-checkbox"
