+++ src/scss/extra-components/_datatable.scss
@@ -53,6 +53,20 @@ table.dataTable.table-hover > tbody > tr:hover > *{
     position: relative;
   }
 
+  &.processing {
+    .table-responsive::before {
+      content: '';
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      background-color: rgba(0, 0, 0, 0.3);
+      backdrop-filter: blur(2px);
+      z-index: 1;
+    }
+  }
+
   .dataTables_processing {
     position: absolute;
     top: 50%;
@@ -65,6 +79,7 @@ table.dataTable.table-hover > tbody > tr:hover > *{
     background-color: $table-loading-message-bg;
     box-shadow: $table-loading-message-box-shadow;
     transform: translateX(-50%) translateY(-50%);
+    z-index: 2;
     @include border-radius($border-radius);
 
     > div {
